web/requirements.txt
fastapi
pydantic
uvicorn
pika
intake
pytest
jinja2
pyjwt==1.7.1
sqlalchemy